Move test cases for error handling
